Jail functionality complete
All of the logic for the jail being handled has been complete. There are only 3 options, either you pay the fine, must roll a 1 to be let out, or use a get out of jail free card if you have one. There is no rule for time passing to be let out of jail.